+++ particle-node/src/node.rs
@@ -52,6 +52,7 @@ use spell_event_bus::bus::SpellEventBus;
 use tokio::sync::{mpsc, oneshot};
 use tokio::task;
 
+use crate::builtins::make_peer_builtin;
 use crate::dispatcher::Dispatcher;
 use crate::effectors::Effectors;
 use crate::metrics::start_metrics_endpoint;
@@ -248,22 +249,23 @@ impl<RT: AquaRuntime> Node<RT> {
         let (spell_event_bus, spell_event_bus_api, spell_events_receiver) =
             SpellEventBus::new(sources);
 
-        let initial_node_info = NodeInfo {
-            external_addresses: config.external_addresses(),
-            node_version: env!("CARGO_PKG_VERSION"),
-            air_version: air_interpreter_wasm::VERSION,
-            spell_version: fluence_spell_distro::VERSION.to_string(),
-            allowed_binaries,
-        };
-        let (sorcerer, spell_service_functions) = Sorcerer::new(
+        let (sorcerer, mut spell_service_functions, spell_version) = Sorcerer::new(
             builtins.services.clone(),
             builtins.modules.clone(),
             aquamarine_api.clone(),
             config.clone(),
             spell_event_bus_api,
             key_manager.clone(),
-            initial_node_info,
         );
+
+        let node_info = NodeInfo {
+            external_addresses: config.external_addresses(),
+            node_version: env!("CARGO_PKG_VERSION"),
+            air_version: air_interpreter_wasm::VERSION,
+            spell_version,
+            allowed_binaries,
+        };
+        spell_service_functions.extend_one(make_peer_builtin(node_info));
 
         spell_service_functions.into_iter().for_each(
             move |(
